fix(lightwell): extract nav outside card + fix masthead stacking
RHCLOUD-50417
Move horizontal subnav from Page horizontalSubnav prop (renders inside
felt-theme card) to a standalone LightwellNavigation component rendered
via Fragment in the masthead prop. Nav wrapper uses grid-column: 1 / -1
to span PF6 Page's CSS grid. Change Masthead display to inline at all
breakpoints so toolbar sits beside title.
